CAR-5804: write stories in .tsx
-- select with autosuggest
