+++ src/stories/select/simple/label.stories.tsx
@@ -1,14 +1,13 @@
-import React, { useState } from "react"
+import React, { useState, FC } from "react"
 import { action } from "@storybook/addon-actions"
 
 import StoryContainer from "../../storyContainer"
 import Select from "../../../components/select"
 
 const handleChange = () => action("handleChange")
-const initialValue = ""
 
 export default {
-  title: "Select  with autosuggest/Appearance",
+  title: "Select/Label",
   component: Select,
   args: {
     storyName: "",
@@ -21,7 +20,8 @@ export default {
       { value: 6, name: "Six" },
     ],
     skipContainer: false,
-    withAutosuggest: true,
+    withAutosuggest: false,
+    labelText: "Label",
   },
   argTypes: {
     buttonName: {
@@ -29,14 +29,20 @@ export default {
         disable: true,
       },
     },
-    skipContainer: {
-      description:
-        "`skipContainer` is useful if you want to have more control over how much the menu spans. You need to put the `<Select>` in a relative position container yourself then.",
-    },
   },
 }
 
-const Template = (args) => {
+interface Props {
+  storyName: string
+  skipContainer: boolean
+  selected: boolean
+  handleChange: (value: string) => void
+  name: string
+  options: Array<{ name: string; value: any | { customValue: any } }>
+  withAutosuggest: boolean
+}
+
+const Template: FC<Props> = ({ ...args } : Props) => {
   const [value, setValue] = useState(null)
   return (
     <StoryContainer
@@ -48,7 +54,7 @@ const Template = (args) => {
               <Select
                 {...args}
                 selected={value}
-                handleChange={(v) => {
+                handleChange={(v: React.ChangeEvent<HTMLSelectElement>) => {
                   setValue(v)
                   handleChange()(v)
                 }}
@@ -59,7 +65,7 @@ const Template = (args) => {
           <Select
             {...args}
             selected={value}
-            handleChange={(v) => {
+            handleChange={(v: React.ChangeEvent<HTMLSelectElement>) => {
               setValue(v)
               handleChange()(v)
             }}
@@ -70,51 +76,44 @@ const Template = (args) => {
   )
 }
 
-export const Simple = Template.bind({})
-Simple.args = {
-  storyName: "Simple",
-}
-
-export const WithSearchIcon = Template.bind({})
-WithSearchIcon.args = {
-  storyName: "With search icon",
-  showSearchIcon: true,
-}
-
-export const Disabled = Template.bind({})
-Disabled.args = {
-  storyName: "Disabled",
-  disabled: true,
-}
-
-export const WithErrorMessage = Template.bind({})
-WithErrorMessage.args = {
-  storyName: "With error message",
-  error: "Error message",
+export const Standard = Template.bind({})
+Standard.args = {
+  storyName: "Standard",
 }
 
 export const WithPlaceholder = Template.bind({})
 WithPlaceholder.args = {
-  storyName: "With placeholder",
   placeholder: "Placeholder",
+  storyName: "With placeholder",
 }
 
 export const WithHint = Template.bind({})
 WithHint.args = {
-  storyName: "With hint",
   hint: "Hint text",
+  storyName: "With hint",
 }
 
-export const WithNoResultsMessage = Template.bind({})
-WithNoResultsMessage.args = {
-  selected: initialValue,
-  noResultsText: "Sorry no results",
-  storyName: "With no results message",
+export const RequiredIndicator = Template.bind({})
+RequiredIndicator.args = {
+  required: true,
+  storyName: "Required indicator",
 }
 
-export const WithCustomWrapper = Template.bind({})
-WithCustomWrapper.args = {
-  selected: initialValue,
-  skipContainer: true,
-  storyName: "With custom wrapper",
+export const Popup = Template.bind({})
+Popup.args = {
+  renderLabelPopup: () => <div>Popup Content</div>,
+  storyName: "Popup",
+}
+
+export const WithErrorMessage = Template.bind({})
+WithErrorMessage.args = {
+  error: "Error message",
+  storyName: "With error message",
+}
+
+export const RequiredWithPopup = Template.bind({})
+RequiredWithPopup.args = {
+  required: true,
+  renderLabelPopup: () => <div>Popup Content</div>,
+  storyName: "Required with Popup",
 }
